Create sample1.fhir.json
The FHIR (Fast Healthcare Interoperability Resources) standard, developed by HL7, is a specification for exchanging healthcare information electronically. The Gravity Project is an initiative that focuses on the standardization of social determinants of health (SDOH) data for interoperability.

In FHIR, SDOH data can be represented using various resources, but one of the most relevant for capturing assessment data is the QuestionnaireResponse resource. This resource is used to record structured data from answers to questions defined in a Questionnaire resource.

Synthetic example of FHIR data in JSON format for two patients' responses to a subset of PRAPARE questionnaire items, following the FHIR standard and considering the Gravity Project's guidelines.